Switch to TypeScript
I originally thought that it would be best if we just used JSDoc
comments, but we were transpiling because of JSX already anyway, and
actually transpiling using TypeScript means we don't need babel. And
having _both_ typescript and babel was really making the list of config
files and dev dependencies long.