Setup storage (#1)
* Setup a storage context

- This is very much a first pass, and I'm going to change things.
- The storage context doesn't currently use local storage, it just has a
  reducer state in memory, but this was just to get the setup working,
  with the idea being that the storage system should be able to hide
  however it actually persists memory from whatever uses it.
- I set up the theme context to use the storage context.
- I did a lot of wrangling in src/index.tsx for this. More than I
  probably should to make arbitrary application nice.
- In light of this, I think storage shouldn't actually be a context. It
  should just be something that other contexts use, but I wanted to
  commit this version so I don't lose it.

* Switch storage to just being a reducer

It will probably need to switch to being a hook once I add the side
effect of actually persisting the state.

But, this greatly reduces the complexity of logic that has to happen at
the top level.

* Switch storage to be a hook

Though at the moment it doesn't do anything more than useReducer.

* Add live-updating of the system theme

Possibly not doesn't belong on this branch in retrospect but there it
is.

* Actually store the storage in local storage

Most of the actual code here though was in validating and serialising
and deserialisng and whatnot. Not sure that we will stick with this
forever, and there are a lot of todos, but I like this for now.

* Log errors on failed storage deserialisation

A better approach will probably be needed in the long run still. The
deserialisation function should probably error and be handled at the
highest level.